requirements.txt
Flask>=3.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
numpy>=1.26.0
face-recognition>=1.3.0
